Factor out message handling.
